Works partially

+++ src/app/dvd.service.ts
@@ -29,25 +29,24 @@ export class DvdService {
     return this.http.get('http://localhost:8080/dvdapi/dvds/rating/' + term);
   }
 
-  // public createDvd(newDvd) {
-  //   return this.http.post('http://localhost:8080/dvdapi/dvd', {
-  //     Data : {
-  //       "id": newDvd.id,
-  //       "title": newDvd.title,
-  //       "releaseYear": newDvd.releaseYear,
-  //       "director": newDvd.director,
-  //       "rating": newDvd.rating,
-  //       "notes": newDvd.notes
-  //     }
-  //   });
-  // }
+  public createDvd(newDvd) {
+    return this.http.post('http://localhost:8080/dvdapi/dvd', {
+        "title": newDvd.dvdData.dvdTitle,
+        "releaseYear": newDvd.dvdData.releaseYear,
+        "director": newDvd.dvdData.director,
+        "rating": newDvd.dvdData.rating,
+        "notes": newDvd.dvdData.notes
+    });
+  }
 
   // public updateDvd(id, dvdData) {
   //   return this.http.put('http://localhost:8080/dvdapi/dvd' + id);
   // }
 
   public deleteDvd(id: number) {
-    return this.http.delete('http://localhost:8080/dvdapi/dvd/' + id);
+    if(confirm("Are you sure you want to delete this DVD from your collection?")) {
+      return this.http.delete('http://localhost:8080/dvdapi/dvd/' + id);
+    }
   }
 
 }
